✨ Feat: ViewModel 통합
